pyproject.toml
[project]
name = "finance-qa-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-openai",
    "langgraph",
    "langgraph-cli[inmem]",
    "langsmith",
    "python-dotenv",
    "scikit-learn",
    "numpy",
]
